refactor: change the type of total_score column
the total_score column in coins table was INTEGER
but happened a problem when calculating the real score
dividing (total_score / number_of_reviews) because it was
returning a rouded value, so the type of total_score
was changed to REAL.

+++ DBConnection.py
@@ -19,7 +19,7 @@ class DBConnection:
                 symbol TEXT,
                 icon TEXT,
                 price TEXT,
-                total_score INTEGER,
+                total_score REAL,
                 number_of_reviews INTEGER
                 );
                 """
